nuttx: drivers: can: fix loop on can_write api when enable canfd
-  fix loop on can_write api when enable canfd

+++ drivers/can/can.c
@@ -761,7 +761,7 @@ static ssize_t can_write(FAR struct file *filep, FAR const char *buffer,
    * shorter than the minimum.
    */
 
-  while (((ssize_t)buflen - nsent) >= CAN_MSGLEN(0))
+  while (((ssize_t)buflen - nsent) >= (ssize_t)CAN_MSGLEN(0))
     {
       /* Check if adding this new message would over-run the drivers ability
        * to enqueue xmit data.
